おにぎりダイアログを機能単位のコンポーネントに分割
onigiri-dialog.tsxが796行と肥大化していたため、星評価・フォームセクション・
写真アップロード欄・Xインポート・詳細表示・編集フォームをそれぞれ独立した
コンポーネントに分離し、ダイアログ本体を状態管理とレイアウト調整のみを担う
コーディネーターに整理した。挙動・テスト結果に変更はない。
