+++ app/components/onigiri/form-section.tsx
@@ -1,0 +1,19 @@
+"use client";
+
+import React from "react";
+
+/**
+ * フォームセクションコンポーネント
+ */
+export function FormSection({ title, children }: { title: string; children: React.ReactNode }) {
+  return (
+    <div className="space-y-3">
+      <h3 className="text-sm font-semibold text-muted-foreground border-b border-border pb-1">
+        {title}
+      </h3>
+      <div className="space-y-3">
+        {children}
+      </div>
+    </div>
+  );
+}
