+++ app/components/onigiri/onigiri-dialog.tsx
@@ -5,51 +5,10 @@ import { Onigiri, CreateOnigiriInput } from "../../models/Onigiri";
 import { formatDisplayDate, formatDateToString } from "../../utils/date-utils";
 import { Button } from "../ui/button";
 import { DialogTitle, DialogDescription } from "../ui/dialog";
-import Image from "next/image";
-import { Plus, Trash2, Link } from "lucide-react";
-import { extractTweetId } from "../../utils/tweet-parser";
-import { deleteImageByUrl, uploadImage, downloadAndUploadImage } from "../../services/image-service";
-
-/**
- * 星評価コンポーネント
- */
-function StarRating({ rating, onRatingChange }: { rating: number; onRatingChange?: (value: number) => void }) {
-  const stars = Array.from({ length: 5 }).map((_, index) => index + 1);
-
-  return (
-    <div className="flex gap-2">
-      {stars.map((star) => (
-        <button
-          key={star}
-          type="button"
-          disabled={!onRatingChange}
-          onClick={() => onRatingChange?.(star)}
-          className={`text-3xl p-1 transition-transform ${
-            star <= rating ? "text-yellow-400" : "text-gray-300"
-          } ${onRatingChange ? "cursor-pointer active:scale-110" : "cursor-default"}`}
-        >
-          ★
-        </button>
-      ))}
-    </div>
-  );
-}
-
-/**
- * フォームセクションコンポーネント
- */
-function FormSection({ title, children }: { title: string; children: React.ReactNode }) {
-  return (
-    <div className="space-y-3">
-      <h3 className="text-sm font-semibold text-muted-foreground border-b border-border pb-1">
-        {title}
-      </h3>
-      <div className="space-y-3">
-        {children}
-      </div>
-    </div>
-  );
-}
+import { Trash2 } from "lucide-react";
+import { deleteImageByUrl, uploadImage } from "../../services/image-service";
+import { OnigiriForm } from "./onigiri-form";
+import { OnigiriDetailView } from "./onigiri-detail-view";
 
 /**
  * おにぎり詳細ダイアログのProps
@@ -89,12 +48,6 @@ export function OnigiriDialog({ isOpen, onClose, date, onigiri, onSave, onDelete
   const [isDeleteConfirming, setIsDeleteConfirming] = useState(false);
   const [isDeleting, setIsDeleting] = useState(false);
 
-  // Xポストインポート状態
-  const [tweetUrl, setTweetUrl] = useState("");
-  const [isFetchingTweet, setIsFetchingTweet] = useState(false);
-  const [tweetFetchError, setTweetFetchError] = useState("");
-  const [tweetFetchSuccess, setTweetFetchSuccess] = useState(false);
-
   // Props変更時の処理
   useEffect(() => {
     setIsEditing(!onigiri);
@@ -113,9 +66,6 @@ export function OnigiriDialog({ isOpen, onClose, date, onigiri, onSave, onDelete
     setTouched({});
     setIsDeleteConfirming(false);
     setIsDeleting(false);
-    setTweetUrl("");
-    setTweetFetchError("");
-    setTweetFetchSuccess(false);
 
     isFirstRender.current = false;
   }, [date, onigiri, isOpen]);
@@ -131,6 +81,9 @@ export function OnigiriDialog({ isOpen, onClose, date, onigiri, onSave, onDelete
     memo: onigiri?.memo || "",
     date: onigiri?.date || formatDateToString(date)
   });
+
+  // Xポストインポートセクションのリマウント用key（ダイアログ再オープン時に状態をリセットする）
+  const tweetImportKey = `${isOpen}-${onigiri?.id ?? "new"}-${formatDateToString(date)}`;
 
   // 画像アップロード処理
   const handleImageUpload = async (e: React.ChangeEvent<HTMLInputElement>, fieldName: "imageUrl" | "eatImageUrl") => {
@@ -172,71 +125,24 @@ export function OnigiriDialog({ isOpen, onClose, date, onigiri, onSave, onDelete
     }
   };
 
-  // Xポストからデータを取得するハンドラ
-  const handleFetchTweet = async (urlOverride?: string) => {
-    const url = urlOverride || tweetUrl;
-    if (!url.trim()) return;
+  // Xポストのテキストデータをフォームに反映（空フィールドのみ）
+  const handleTweetTextFetched = ({ text, parsed }: { text: string; parsed: { name?: string; storeName?: string; price?: number } }) => {
+    setFormData((prev) => ({
+      ...prev,
+      name: prev.name || parsed.name || "",
+      storeName: prev.storeName || parsed.storeName || "",
+      price: prev.price ?? parsed.price ?? 0,
+      memo: prev.memo || text || "",
+    }));
+  };
 
-    if (!extractTweetId(url)) {
-      setTweetFetchError("有効なXのポストURLを入力してください");
-      return;
-    }
-
-    setIsFetchingTweet(true);
-    setTweetFetchError("");
-    setTweetFetchSuccess(false);
-
-    try {
-      const response = await fetch("/api/fetch-tweet", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ url }),
-      });
-
-      if (!response.ok) {
-        const errorData = await response.json();
-        setTweetFetchError(errorData.error || "ポストのデータを取得できませんでした");
-        return;
-      }
-
-      const data = await response.json();
-
-      // テキストデータでフォームを更新（空フィールドのみ）
-      setFormData((prev) => ({
-        ...prev,
-        name: prev.name || data.parsed.name || "",
-        storeName: prev.storeName || data.parsed.storeName || "",
-        price: prev.price ?? data.parsed.price ?? 0,
-        memo: prev.memo || data.text || "",
-      }));
-
-      // 画像をSupabaseにダウンロード・アップロード（2枚並行）
-      if (data.images.length >= 2) {
-        try {
-          setIsImageUploading(true);
-          const [imagePublicUrl, eatImagePublicUrl] = await Promise.all([
-            downloadAndUploadImage(data.images[0]),
-            downloadAndUploadImage(data.images[1]),
-          ]);
-          setFormData((prev) => ({
-            ...prev,
-            imageUrl: imagePublicUrl,
-            eatImageUrl: eatImagePublicUrl,
-          }));
-        } catch (imgError) {
-          console.warn("画像のダウンロードに失敗:", imgError);
-        } finally {
-          setIsImageUploading(false);
-        }
-      }
-
-      setTweetFetchSuccess(true);
-      setTimeout(() => setTweetFetchSuccess(false), 3000);
-    } catch {
-      setTweetFetchError("通信エラーが発生しました。もう一度お試しください。");
-    } finally {
-      setIsFetchingTweet(false);
-    }
+  // Xポストの画像をフォームに反映
+  const handleTweetImagesUploaded = (imageUrl: string, eatImageUrl: string) => {
+    setFormData((prev) => ({
+      ...prev,
+      imageUrl,
+      eatImageUrl,
+    }));
   };
 
   // フォーム入力の変更ハンドラ
@@ -276,7 +182,7 @@ export function OnigiriDialog({ isOpen, onClose, date, onigiri, onSave, onDelete
     console.log('おにぎりを保存します:', formData);
   };
 
-  // ダイアログを閉じる処理
+  // ダイアログを閉じる処理（Radix Dialog向け、未使用時の将来的な拡張のため保持）
   const handleDialogClose = (open: boolean) => {
     if (!open) {
       onClose();
@@ -348,386 +254,27 @@ export function OnigiriDialog({ isOpen, onClose, date, onigiri, onSave, onDelete
       <div className="p-3 overflow-y-auto flex-1 bg-card">
         {isEditing ? (
           // 編集フォーム
-          <form className="space-y-5 py-2">
-            {/* セクション: Xからインポート */}
-            <FormSection title="Xからインポート">
-              <div className="flex gap-2">
-                <div className="relative flex-1">
-                  <Link className="absolute left-2.5 top-1/2 -translate-y-1/2 h-4 w-4 text-muted-foreground" />
-                  <input
-                    type="text"
-                    value={tweetUrl}
-                    onChange={(e) => {
-                      setTweetUrl(e.target.value);
-                      setTweetFetchError("");
-                    }}
-                    onPaste={(e) => {
-                      const pasted = e.clipboardData.getData("text");
-                      if (extractTweetId(pasted)) {
-                        e.preventDefault();
-                        setTweetUrl(pasted);
-                        setTimeout(() => handleFetchTweet(pasted), 0);
-                      }
-                    }}
-                    placeholder="https://x.com/user/status/..."
-                    className="w-full p-2 pl-8 border border-input rounded-md bg-background text-foreground text-sm"
-                  />
-                </div>
-                <Button
-                  type="button"
-                  variant="outline"
-                  onClick={() => handleFetchTweet()}
-                  disabled={isFetchingTweet || !tweetUrl.trim()}
-                  className="px-3 py-1 whitespace-nowrap"
-                >
-                  {isFetchingTweet ? (
-                    <div className="animate-spin rounded-full h-4 w-4 border-t-2 border-b-2 border-orange-500" />
-                  ) : (
-                    "取得"
-                  )}
-                </Button>
-              </div>
-              {tweetFetchError && (
-                <p className="text-xs text-red-500">{tweetFetchError}</p>
-              )}
-              {tweetFetchSuccess && (
-                <p className="text-xs text-green-600">ポストからデータを取得しました</p>
-              )}
-            </FormSection>
-
-            {/* セクション: 基本情報 */}
-            <FormSection title="基本情報">
-              <div className="grid grid-cols-1 gap-3 sm:grid-cols-2 sm:gap-4">
-                <div>
-                  <label htmlFor="name" className="block text-sm font-medium text-foreground mb-1">
-                    おにぎり名 <span className="text-red-500">*</span>
-                  </label>
-                  <input
-                    type="text"
-                    id="name"
-                    name="name"
-                    required
-                    value={formData.name}
-                    onChange={handleChange}
-                    onBlur={handleBlur}
-                    className={`w-full p-2 border rounded-md bg-background text-foreground ${
-                      touched.name && !formData.name
-                        ? "border-red-500 ring-1 ring-red-500"
-                        : "border-input"
-                    }`}
-                  />
-                  {touched.name && !formData.name && (
-                    <p className="text-xs text-red-500 mt-1">おにぎり名は必須です</p>
-                  )}
-                </div>
-
-                <div>
-                  <label htmlFor="storeName" className="block text-sm font-medium text-foreground mb-1">
-                    店舗名 <span className="text-red-500">*</span>
-                  </label>
-                  <input
-                    type="text"
-                    id="storeName"
-                    name="storeName"
-                    required
-                    value={formData.storeName}
-                    onChange={handleChange}
-                    onBlur={handleBlur}
-                    className={`w-full p-2 border rounded-md bg-background text-foreground ${
-                      touched.storeName && !formData.storeName
-                        ? "border-red-500 ring-1 ring-red-500"
-                        : "border-input"
-                    }`}
-                  />
-                  {touched.storeName && !formData.storeName && (
-                    <p className="text-xs text-red-500 mt-1">店舗名は必須です</p>
-                  )}
-                </div>
-
-                <div>
-                  <label htmlFor="price" className="block text-sm font-medium text-foreground mb-1">
-                    価格 <span className="text-red-500">*</span>
-                  </label>
-                  <div className="relative">
-                    <input
-                      type="number"
-                      id="price"
-                      name="price"
-                      min="0"
-                      value={formData.price === null ? "" : formData.price}
-                      onChange={handleChange}
-                      onBlur={handleBlur}
-                      className="w-full p-2 pr-8 border border-input rounded-md bg-background text-foreground"
-                    />
-                    <span className="absolute right-3 top-1/2 -translate-y-1/2 text-sm text-muted-foreground pointer-events-none">
-                      円
-                    </span>
-                  </div>
-                </div>
-
-                <div>
-                  <label htmlFor="rating" className="block text-sm font-medium text-foreground mb-1">
-                    評価
-                  </label>
-                  <StarRating rating={formData.rating} onRatingChange={handleRatingChange} />
-                </div>
-              </div>
-            </FormSection>
-
-            {/* セクション: 写真 */}
-            <FormSection title="写真">
-              <div className="grid grid-cols-2 gap-3">
-                {/* おにぎりの写真 */}
-                <div className="space-y-1">
-                  <label className="block text-sm font-medium text-foreground mb-1">
-                    おにぎりの写真
-                  </label>
-                  <label
-                    htmlFor="imageUpload"
-                    className={`relative flex flex-col items-center justify-center w-full h-32 rounded-xl border-2 border-dashed cursor-pointer transition-colors ${
-                      formData.imageUrl
-                        ? "border-solid border-border"
-                        : "border-input bg-muted/30 hover:bg-muted/50 hover:border-primary/50"
-                    }`}
-                  >
-                    <input
-                      type="file"
-                      id="imageUpload"
-                      accept="image/*"
-                      onChange={(e) => handleImageUpload(e, "imageUrl")}
-                      className="hidden"
-                    />
-                    {formData.imageUrl ? (
-                      <Image
-                        src={formData.imageUrl}
-                        alt="プレビュー"
-                        fill
-                        sizes="(max-width: 500px) 50vw, 250px"
-                        className="object-cover rounded-xl"
-                        onError={(e) => {
-                          // @ts-ignore
-                          e.currentTarget.src = "/images/onigiri-sample-1.jpg";
-                          // @ts-ignore
-                          e.currentTarget.onerror = null;
-                        }}
-                      />
-                    ) : (
-                      <div className="flex flex-col items-center gap-1 text-muted-foreground">
-                        <Plus className="h-8 w-8" />
-                        <span className="text-xs">写真を追加</span>
-                      </div>
-                    )}
-                    {isImageUploading && (
-                      <div className="absolute inset-0 flex items-center justify-center bg-card/80 rounded-xl">
-                        <div className="animate-spin rounded-full h-8 w-8 border-t-2 border-b-2 border-orange-500"></div>
-                      </div>
-                    )}
-                  </label>
-                  <button
-                    type="button"
-                    onClick={() => setShowImageUrlInput(!showImageUrlInput)}
-                    className="text-xs text-muted-foreground underline hover:text-foreground"
-                  >
-                    {showImageUrlInput ? "URL入力を閉じる" : "URLを直接入力する"}
-                  </button>
-                  {showImageUrlInput && (
-                    <input
-                      type="text"
-                      id="imageUrl"
-                      name="imageUrl"
-                      value={formData.imageUrl || ""}
-                      onChange={handleChange}
-                      className="w-full p-2 border border-input rounded-md bg-background text-foreground text-xs"
-                      placeholder="https://example.com/image.jpg"
-                    />
-                  )}
-                </div>
-
-                {/* 食べた時の写真 */}
-                <div className="space-y-1">
-                  <label className="block text-sm font-medium text-foreground mb-1">
-                    食べた時の写真
-                  </label>
-                  <label
-                    htmlFor="eatImageUpload"
-                    className={`relative flex flex-col items-center justify-center w-full h-32 rounded-xl border-2 border-dashed cursor-pointer transition-colors ${
-                      formData.eatImageUrl
-                        ? "border-solid border-border"
-                        : "border-input bg-muted/30 hover:bg-muted/50 hover:border-primary/50"
-                    }`}
-                  >
-                    <input
-                      type="file"
-                      id="eatImageUpload"
-                      accept="image/*"
-                      onChange={(e) => handleImageUpload(e, "eatImageUrl")}
-                      className="hidden"
-                    />
-                    {formData.eatImageUrl ? (
-                      <Image
-                        src={formData.eatImageUrl}
-                        alt="プレビュー"
-                        fill
-                        sizes="(max-width: 500px) 50vw, 250px"
-                        className="object-cover rounded-xl"
-                        onError={(e) => {
-                          // @ts-ignore
-                          e.currentTarget.src = "/images/onigiri-sample-1.jpg";
-                          // @ts-ignore
-                          e.currentTarget.onerror = null;
-                        }}
-                      />
-                    ) : (
-                      <div className="flex flex-col items-center gap-1 text-muted-foreground">
-                        <Plus className="h-8 w-8" />
-                        <span className="text-xs">写真を追加</span>
-                      </div>
-                    )}
-                    {isImageUploading && (
-                      <div className="absolute inset-0 flex items-center justify-center bg-card/80 rounded-xl">
-                        <div className="animate-spin rounded-full h-8 w-8 border-t-2 border-b-2 border-orange-500"></div>
-                      </div>
-                    )}
-                  </label>
-                  <button
-                    type="button"
-                    onClick={() => setShowEatImageUrlInput(!showEatImageUrlInput)}
-                    className="text-xs text-muted-foreground underline hover:text-foreground"
-                  >
-                    {showEatImageUrlInput ? "URL入力を閉じる" : "URLを直接入力する"}
-                  </button>
-                  {showEatImageUrlInput && (
-                    <input
-                      type="text"
-                      id="eatImageUrl"
-                      name="eatImageUrl"
-                      value={formData.eatImageUrl || ""}
-                      onChange={handleChange}
-                      className="w-full p-2 border border-input rounded-md bg-background text-foreground text-xs"
-                      placeholder="https://example.com/image.jpg"
-                    />
-                  )}
-                </div>
-              </div>
-              {imageUploadError && (
-                <p className="text-xs text-red-500">{imageUploadError}</p>
-              )}
-            </FormSection>
-
-            {/* セクション: メモ */}
-            <FormSection title="メモ">
-              <div>
-                <textarea
-                  id="memo"
-                  name="memo"
-                  rows={2}
-                  value={formData.memo || ""}
-                  onChange={handleChange}
-                  className="w-full p-2 border border-input rounded-md bg-background text-foreground"
-                  placeholder="おにぎりの感想やメモを入力..."
-                />
-              </div>
-            </FormSection>
-          </form>
+          <OnigiriForm
+            formData={formData}
+            touched={touched}
+            isImageUploading={isImageUploading}
+            imageUploadError={imageUploadError}
+            showImageUrlInput={showImageUrlInput}
+            showEatImageUrlInput={showEatImageUrlInput}
+            tweetImportKey={tweetImportKey}
+            onChange={handleChange}
+            onBlur={handleBlur}
+            onRatingChange={handleRatingChange}
+            onImageUpload={handleImageUpload}
+            onToggleImageUrlInput={() => setShowImageUrlInput(!showImageUrlInput)}
+            onToggleEatImageUrlInput={() => setShowEatImageUrlInput(!showEatImageUrlInput)}
+            onTweetTextFetched={handleTweetTextFetched}
+            onTweetImagesUploaded={handleTweetImagesUploaded}
+            onTweetUploadingChange={setIsImageUploading}
+          />
         ) : onigiri ? (
           // 詳細表示
-          <div className="space-y-4 py-2">
-            {/* ヒーロー画像 */}
-            {onigiri.imageUrl && (
-              <div className="relative w-full h-48 -mx-3 -mt-2 overflow-hidden" style={{ width: 'calc(100% + 1.5rem)' }}>
-                <Image
-                  src={onigiri.imageUrl}
-                  alt={onigiri.name}
-                  fill
-                  sizes="(max-width: 768px) 100vw, 500px"
-                  className="object-cover"
-                  onError={(e) => {
-                    // @ts-ignore
-                    e.currentTarget.src = "/images/onigiri-sample-1.jpg";
-                    // @ts-ignore
-                    e.currentTarget.onerror = null;
-                  }}
-                />
-                <div className="absolute inset-0 bg-gradient-to-t from-card/80 to-transparent" />
-                <h2 className="absolute bottom-3 left-4 text-xl font-bold text-white drop-shadow-md">
-                  {onigiri.name}
-                </h2>
-              </div>
-            )}
-
-            {/* 情報カード */}
-            <div className="grid grid-cols-3 gap-2">
-              <div className="bg-muted/50 rounded-lg p-3 text-center">
-                <p className="text-xs text-muted-foreground">店舗</p>
-                <p className="text-sm font-medium text-foreground truncate">{onigiri.storeName}</p>
-              </div>
-              <div className="bg-muted/50 rounded-lg p-3 text-center">
-                <p className="text-xs text-muted-foreground">価格</p>
-                <p className="text-sm font-medium text-foreground">{onigiri.price}円</p>
-              </div>
-              <div className="bg-muted/50 rounded-lg p-3 text-center">
-                <p className="text-xs text-muted-foreground">評価</p>
-                <div className="flex justify-center mt-0.5">
-                  <StarRating rating={onigiri.rating} />
-                </div>
-              </div>
-            </div>
-
-            {/* 写真比較 */}
-            {(onigiri.imageUrl || onigiri.eatImageUrl) && (
-              <div className="grid grid-cols-2 gap-2">
-                {onigiri.imageUrl && (
-                  <div>
-                    <p className="text-xs text-muted-foreground mb-1">おにぎり</p>
-                    <div className="relative w-full h-32 rounded-lg overflow-hidden">
-                      <Image
-                        src={onigiri.imageUrl}
-                        alt={onigiri.name}
-                        fill
-                        sizes="(max-width: 500px) 50vw, 250px"
-                        className="object-cover"
-                        onError={(e) => {
-                          // @ts-ignore
-                          e.currentTarget.src = "/images/onigiri-sample-1.jpg";
-                          // @ts-ignore
-                          e.currentTarget.onerror = null;
-                        }}
-                      />
-                    </div>
-                  </div>
-                )}
-
-                {onigiri.eatImageUrl && (
-                  <div>
-                    <p className="text-xs text-muted-foreground mb-1">食べた写真</p>
-                    <div className="relative w-full h-32 rounded-lg overflow-hidden">
-                      <Image
-                        src={onigiri.eatImageUrl}
-                        alt={`${onigiri.name}を食べたところ`}
-                        fill
-                        sizes="(max-width: 500px) 50vw, 250px"
-                        className="object-cover"
-                        onError={(e) => {
-                          // @ts-ignore
-                          e.currentTarget.src = "/images/onigiri-sample-1.jpg";
-                          // @ts-ignore
-                          e.currentTarget.onerror = null;
-                        }}
-                      />
-                    </div>
-                  </div>
-                )}
-              </div>
-            )}
-
-            {/* メモ */}
-            {onigiri.memo && (
-              <div>
-                <h3 className="text-sm font-medium text-foreground">メモ</h3>
-                <p className="mt-1 text-sm text-foreground whitespace-pre-line">{onigiri.memo}</p>
-              </div>
-            )}
-          </div>
+          <OnigiriDetailView onigiri={onigiri} />
         ) : null}
       </div>
